restructure monitoring components
